news controller, staff members controller and testimonies controller updated

+++ controllers/staffController.js
@@ -30,43 +30,29 @@ exports.get_all_staff_members = (req, res) => {
 };
 
 exports.get_staff_member_by_id = async (req, res) => {
-  uid = req.params.id;
-  const staff = await StaffMember.findByPk(uid);
-  if (staff === null) {
-    res
-      .status(404)
-      .json(
-        response.failure_response(
-          errorMessage.error_404("Staff Member"),
-          404,
-          null
-        )
-      );
+  const uid = req.params.id;
+  const convertedToNum = parseInt(uid);
+  if (isNaN(convertedToNum)) {
+    res.status(400).json({
+      message: 'the parameter is not an integer, please provide the actual integer value for the id',
+    })
+    return
+  }
+  if (!uid) {
+    res.status(404).json({
+      status: false,
+      code: 404,
+      message: "staff member is not found",
+    });
     return;
   }
-  staff
-    .then((staff_member) => {
-      res
-        .status(200)
-        .json(
-          response.success_response(
-            sucessMessage.found_by_id_message("Staff Member"),
-            200,
-            staff_member
-          )
-        );
-    })
-    .catch((err) => {
-      res
-        .status(500)
-        .json(
-          response.failure_response(
-            errorMessage.found_by_id_error("Staff Members", uid),
-            500,
-            err
-          )
-        );
-    });
+  const staff_member = await StaffMember.findByPk(convertedToNum);
+  if (!staff_member) {
+    res.status(404).json(response.failure_response(errorMessage.error_404("Staff Member"), 404, null))
+    return;
+  }
+  res.status(200).json(response.success_response(sucessMessage.found_by_id_message('Staff Member', convertedToNum), 200, staff_member))
+  return;
 };
 
 exports.add_staff_member = async (req, res) => {
@@ -82,75 +68,38 @@ exports.add_staff_member = async (req, res) => {
     imageUrl: req.body.imageUrl,
   });
 
-  newStaff
-    .then((createdStaff) => {
-      res
-        .status(201)
-        .json(
-          response.success_response(
-            sucessMessage.updated_message("Staff Member"),
-            201,
-            createdStaff
-          )
-        );
-    })
-    .catch((err) => {
-      res
-        .status(500)
-        .json(
-          response.failure_response(
-            errorMessage.added_error("Staff Members"),
-            500,
-            err
-          )
-        );
-    });
-};
-
-exports.update_staff_member = (req, res) => {
-  const updatedStaff = {};
-  for (const ops of req.body) {
-    updatedStaff[ops.propName] = ops.value;
+  if (newStaff) {
+    res
+      .status(201)
+      .json(
+        response.success_response(
+          sucessMessage.updated_message("Staff Member"),
+          201,
+          newStaff
+        )
+      );
+  } else {
+    res
+      .status(500)
+      .json(
+        response.failure_response(
+          errorMessage.added_error("Staff Members"),
+          500,
+          err
+        )
+      );
   }
-  StaffMember.update({ _id: req.params.id }, { $set: updatedStaff })
-    .exec()
-    .then((staff) => {
-      res
-        .status(200)
-        .json(
-          response.success_response(
-            sucessMessage.updated_message("Staff Member"),
-            200,
-            staff
-          )
-        );
-    })
-    .catch((err) => {
-      res.status(500).json({
-        status: false,
-        code: 500,
-        message: `error found ${err}`,
-      });
-    });
-  // const updatedStaff = StaffMember.update({
-  //     name: req.body.name,
-  //     gender: req.body.gender,
-  //     serviceTitle: req.body.serviceTitle,
-  //     place: req.body.place,
-  //     ethiopianStaff: req.body.ethiopianStaff,
-  //     whoIsHe: req.body.whoIsHe,
-  //     responsibility: req.body.responsibility,
-  //     passion: req.body.passion,
-  //     imageUrl: req.body.imageUrl,
-  // }, where({
-  //     _id: req.body.id
-  // }))
-
-  //   updatedStaff;
 };
 
-exports.delete_staff_member = async (req, res) => {
+exports.update_staff_member = async (req, res) => {
   const uid = req.params.id;
+  const convertedToNum = parseInt(uid);
+  if (isNaN(convertedToNum)) {
+    res.status(400).json({
+      message: 'the parameter is not an integer, please provide the actual integer value for the id',
+    })
+    return
+  }
   if (!uid) {
     res.status(404).json({
       status: false,
@@ -159,25 +108,88 @@ exports.delete_staff_member = async (req, res) => {
     });
     return;
   }
-
-  const staff_member = await StaffMember.findByPk(uid);
-  if(!staff_member) {
+  const staff_member = await StaffMember.findByPk(convertedToNum);
+  if (!staff_member) {
     res.status(404).json(response.failure_response(errorMessage.error_404("Staff Member"), 404, null))
+    return;
   }
-  
+  const update_staff_member = {};
+  const inputValue = req.body;
+  const keys = Object.keys(inputValue);
+
+  keys.forEach(items => {
+    update_staff_member[items] = inputValue[items]
+  })
+  StaffMember.update({
+    name: update_staff_member.name,
+    gender: update_staff_member.gender,
+    serviceTitle: update_staff_member.serviceTitle,
+    place: update_staff_member.place,
+    ethiopianStaff: update_staff_member.ethiopianStaff,
+    whoIsHe: update_staff_member.whoIsHe,
+    responsibility: update_staff_member.responsibility,
+    passion: update_staff_member.passion,
+    imageUrl: update_staff_member.imageUrl,
+  }, {
+    where: {
+      id: convertedToNum,
+    }
+  }).then((staffmember) => {
+    res
+      .status(201)
+      .json(
+        response.success_response(
+          sucessMessage.updated_message("Staff Member"),
+          201,
+          update_staff_member
+        )
+      );
+  }).catch((err) => {
+    res.status(500).json({
+      status: false,
+      code: 500,
+      message: `error found ${err}`,
+    });
+  });
+};
+
+exports.delete_staff_member = async (req, res) => {
+  const uid = req.params.id;
+  const convertedToNum = parseInt(uid);
+  if (isNaN(convertedToNum)) {
+    res.status(400).json({
+      error: 'the parameter is not an integer, please provide the actual integer value for the id',
+    })
+    return
+  }
+  if (!uid) {
+    res.status(404).json({
+      status: false,
+      code: 404,
+      error: "staff member is not found",
+    });
+    return;
+  }
+
+  const staff_member = await StaffMember.findByPk(convertedToNum);
+  if (!staff_member) {
+    res.status(404).json(response.failure_response(errorMessage.error_404("Staff Member"), 404, null))
+    return;
+  }
+
   StaffMember.destroy({
     where: {
-      id: uid,
+      id: convertedToNum,
     },
   })
-    .then((staff) => {
+    .then((staffValue) => {
       res
         .status(200)
         .json(
           response.success_response(
             sucessMessage.deleted_message("Staff Member"),
             200,
-            staff
+            staff_member
           )
         );
     })
